+++ CHANGELOG.md
@@ -4,6 +4,10 @@ Backup your messages from the Telegram messenger.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Group call message actions
+
 ## [0.1.15] - 2021-07-24
 ### Added
 - Support phone call action